CloudOCR/cloudocr-app/requirements.txt
requests
azure-cognitiveservices-vision-computervision
msrest
openai
fastapi
uvicorn
python-multipart